+++ classes/output/navigation/primary.php
@@ -184,6 +184,16 @@ class primary extends \core\navigation\output\primary {
             if (isset($menu->submenuid)) {
                 $children = $menu->children;
 
+                // Add the second level menus list before the course list to the user menu.
+                // This will have the effect that, when opening the third level submenus, the transition will go to the right.
+                $submenu = [
+                    'id' => $menu->submenuid,
+                    'title' => $menu->title ?: $menu->text,
+                ];
+                $usermenu['submenus'][] = (object) $submenu;
+                // The key of this submenu which helps later to include its children after including the necessary data.
+                $lastkey = array_key_last($usermenu['submenus']);
+
                 // Update the dividers item type.
                 array_walk($children, function(&$value) use (&$usermenu, $menu) {
                     if (isset($value['divider'])) {
@@ -202,6 +212,7 @@ class primary extends \core\navigation\output\primary {
                             'id' => $uniqueid,
                             'returnid' => $menu->submenuid, // Return the third level submenus back to its parent section.
                             'title' => $value['title'],
+                            'text' => strip_tags($value['text']), // Remove the item icon from the submenus title.
                             'items' => $value['children'],
                         ];
 
@@ -212,13 +223,9 @@ class primary extends \core\navigation\output\primary {
                     }
                 });
 
-                $submenu = [
-                    'id' => $menu->submenuid,
-                    'title' => $menu->title,
-                    'items' => $children,
-                ];
+                // Update the children elements for the submenu.
+                $usermenu['submenus'][$lastkey]->items = $children;
                 $usermenu['items'][] = $menu;
-                $usermenu['submenus'][] = (object) $submenu;
             }
         }